pyproject.toml
[project]
name = "iam"
version = "0.1.0"
description = ""
authors = [
    {name = "Suender Oliveira",email = "suender@live.com"}
]
license = {text = "APACHE 2.0"}
readme = "README.md"
package-mode = false
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi[all] (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "redis (>=6.2.0,<7.0.0)",
    "aio-pika (>=9.5.5,<10.0.0)",
    "pybreaker (>=1.3.0,<2.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-asyncio (>=1.0.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
